AHCI: Some cleanup
Until now there were two different methods to forward new requests from GC to R3.
For sync I/O a FIFO and a PDM queue to kick the I/O thread was used.
For async I/O every task was passed through the PDM queue and processed by the consumer.
This is messy and requires too much hyper heap.

The new method is the same for sync and async I/O. Because a port can only have 32 active requests
at a time it is enough to use a bitmap which is processed in R3 to get the slots of the new tasks.
Requires a PDM queue with much less items then before and saves us almost 20KB of hyper heap on a 64bit host.



git-svn-id: http://www.virtualbox.org/svn/vbox@33209 cfe28804-0f27-0410-a406-dd0f0b0b656f