Additions/linux/drm: better detection for EL7 variants. bugref:4567: Linux kernel driver maintenance
r118292/r68647 was very invasive and broke Debian 9 on top.  This change
refines the check for supported guests for the drm driver by checking the
kernel version string to be sure that it either starts with a supported
version (that is, not 3.10.x or older) or contains the EL7 signature.

git-svn-id: http://www.virtualbox.org/svn/vbox@69563 cfe28804-0f27-0410-a406-dd0f0b0b656f