Main/Event: defunct passive event listeners prevent subsequent vetoable events from being processed bugref:10000
If a passive event listener has registered an interest in a vetoable 
event and subsequently disappears then any later passive event listeners
which are interested in the same vetoable event won't ever see
IEvent::setProcessed() called.  The most common scenario where this is
seen is when the VBox GUI crashes and restarts during the same VBoxSVC
instance after which any time an OnExtraDataCanChange event is
triggered, such as by clicking a VM's name, it incurs a three second 
timeout where nothing can be done in the GUI during that time.  The fix
removes OnExtraDataCanChange from the unresponsive passive event
listener's list of interested events after it times out the first time
which means that the annoying three seconds of GUI inactivity will only
be seen once.


git-svn-id: http://www.virtualbox.org/svn/vbox@93188 cfe28804-0f27-0410-a406-dd0f0b0b656f